agent/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents",
    "livekit-plugins-cartesia",
    "livekit-plugins-deepgram",
    "livekit-plugins-openai",
    "livekit-plugins-silero",
    "livekit-plugins-turn-detector",
    "python-dotenv",
    "requests",
]
